Update reference data for record in list test
The new JSON reader has a subtle difference to the old one, which may be
considered a bug fix. The table slice builder API requires us to add
missing fields as `nil` for records, but it doesn't strictly require us
to do this for missing fields in records inside lists. When updating the
JSON reader to work with the new type system I have accidentally aligned
the behavior, so now records inside lists are always expanded. It's
probably better this way, so this commit just updates the test reference
data.