Simplify data extractors
I noticed that outside of unit tests, the type resolver always operated
on record types directly. This allows for a drastic simplification of
the data extractors, which now no longer hold an offset but rather the
calculated flat column index directly, avoiding the need for repetitive
`record_type::flat_index` calls when filtering table slices or selecting
an indexer for a given type.