Avoid evaluating calls when assertions are disabled
Previous to this change the statement `VAST_ASSERT(expr)` would
still evaluate `expr` unless the optimizer was able to eliminate
the code after inlining.

Of course it is better not to rely on that mechanism, and indeed it
sometimes fails even with LTO turned on.

In a simple ingest test on my T14s the suricata parser performance
increased by about 16% and exporting with a highly selective
query performed a whopping 4 times faster.