Remove type name hashing
The legacy type system was the only place that dependend on the hashes
of the type name annotation in inspect overloads to be incldued in the
hash, and now that we don't have that anymore we can remove it. This
should result in a small speedup of hash functions using the hash
inspector and make the branch predictor be in a happier place.