Implement a wrapper for FlatBuffers table pointers
This should help simplify our code a fair bit down the line: A wrapper
around a FlatBuffers table pointer that shares the reference count of
the chunk it is constructed from, and is thus easier to handle via RAII.
A `vast::flatbuffer<T>`  can optionally be constructed from a
`flatbuffers::DetachedBuffer&&`, allowing the developer to skip handling
the chunk entirely. The major advantage of this compared to using a
`vast::chunk` directly is that it is safe to slice a non-root table from
a `vast::flatbuffer`, which helps reduce unnecessary pointer chasing.