Improve caching behavior in CI
GitHub recently increased the cache to 10GB per repository, which caused
me to look at our CI configuration in that regard, where I noticed that
we only accounted for caching for one branch at the same time. This
should improve the situation a fair bit. The inline comment explains the
reasoning behind the chosen size.

Additionally, this takes advantage of us using the modified cache action
that always uploads the cache by changing by making the key unique for
every branch, which in turn makes it so there only is a single cache per
pull request.