Add a coroutine-based generator type
This takes `generator<T>` from the popular cppcoro library from Lewis
Baker, and makes use if it in the new type system in places where
previously we either returned a custom proxy type that had a
range-interface or a vector. The resulting code changes for
`record_type::fields()` and `record_type::leaves()` shows just how big
of a readability improvement this is.