Optimize `record_type::{num_leaves,flat_index}`
This removes the recursive `flat_size` helper function that
`record_type:.num_leaves` and `record_type::flat_index` relied upon
internally and removes it. Instead, I've rewritten the functions to
operate without recursion directly on the underlying FlatBuffers table,
which is a fair bit faster. Considering we spend over 10% of the time
during `export null` in `record_type::flat_index` I expect this to be a
meaningful optimization (measurements yet to be done).