Fix bug creating invalid table slices in the JSON reader
Given this schema and data, we created invalid table slices due to a
mismatch in the amount of fields we create.

Note that while techncially some of the fields could be inferred from
the data, it is more sensible to add `nil` for all of them as the
overall record contains a type mismatch. The `id` field takes precedence
over flattened fields, and that has to stay.

```
type zeek = record {
  _path: string,
  id: record {
    orig_h: addr,
    orig_p: port,
    resp_h: addr,
    resp_p: port,
  },
}
```

```json
{"_path":"conn","id.orig_h":"45.58.146.42","id.orig_p":51472,"id.resp_h":"141.9.217.223","id":0}
```