Remove defunct server heartbeat message code
The periodic throughput status message to STDOUT has been broken
since we switched to spdlog for logging. We never heard a complaint
about that so it's save to remove that logic now.

The status command can be used as an alternative way to make sure
the system is operating correctly.