Always unshare table slices in unit tests
Many unit tests use table slices that are present in the form of
static test data and shared over many tests.

That is a problem in combination with the importer, who expects to
receive unique handles to its table slices in order to be able to
safely modify them.

As a workaround, always make copies of table slices sent using
`detail::spawn_container_source()`.