Add a `--omit-nulls` option to the JSON export
This adds the `vast.export.json.omit-nulls` option that changes the JSON
writer to skip over fields in JSON objects whose value is `null`.

Implementation-wise this turned out to be a bit more complicated than I
had anticipated, as it required changes to both the JSON printer and the
ostream writer. I re-used the same policy template argument for both.