Rebuild older Arrow table slices in `as_record_batch`
This is a bug fix for transforms, which operated under the assumption
that it was always safe to create a new table slice from a record
batch. If, however, assumptions about the record batch memory layout
change between Arrow table slice encoding versions, this will no longer
work, requiring us to rebuild the slice to the latest encoding version
before retrieving the underlying record batch.