Introduce an experimental Arrow table slice encoding
The `experimental` format is explicitly unstable, and serves as a
playground for upcoming changes to the Arrow encoding. We plan to merge
these as a v2 of the current encoding some time in the future, but want
to be able to make incremental changes. Please do not use this feature
in production, as we explicitly do not guarantee backwards compatibility
for it.

To enable permanently for local development, put this into your
`vast.yaml` configuration file:

```yaml
vast:
  import:
    batch-encoding: experimental
```