Make use of `table_slice::append_column_to_index`
Instead of decoding columns in table slice for every row, we should make
use of the existing API for appending a whole column at once. I'm not
sure when this got thrown out, but it's been dead code before this
change, although we definitely do want to make use of it.

An even better API for this would be to not have a custom function in
the table slice API that takes a value index, but rather to return a
`detail::generator<data_view>` per column that can be lazily iterated
over. That still shouldn't stop us from using this special function
before we have a more generic solution, though.