Make opt-in to formatting via deep_to_string easier
This makes it easier for any given struct to be opted into using
`caf::deep_to_string` for formatting with {fmt} by simply defining a static
member variable `use_deep_to_string_formatter` with the value `true`.

As a small bonus, this adds a `use_name_member_formatter` that works similarly
and makes use of the member function `T::name()` if
`T::use_name_member_formatter` is `true`, and makes use of it for event based
actors.