Add FlatBuffers for `data`
This only adds `pack` and `unpack` functions; it doesn't rewrite `data` to
actually dispatch based on the FlatBuffers tables contents, although that'd
probably be a really good optimization to make and certainly would be more
efficient than the `caf::variant` based implementation.

We ultimately need this for the `hash_index` FlatBuffers, which stores `data` as
keys in a map. I initially thought I could get by with using MessagePack for the
encoding, but that is (1) a much larger refactoring than this and (2) requires
knowing _which_ data type you want to decode ahead of time, which would then
need to be separately versioned. The second alternative was using the string
representation, but that is currently not versioned and using that would lock
that representation in place, which is not a decision that should be made
lightly.