Simplify naming scheme in `type.fbs`
The general idea here is that instead of using a _vX suffix, we name the union
members explicitly, adding a suffix for outdated versions only.

The version of FlatBuffers we depended on when writing most of our FlatBuffers
table did not allow for explicitly naming union members, which is part of why we
ended up with the existing naming scheme in the first place.