Add an integration test using our schema to its fullest
We want to have a single integration test that is run against both the
previous table slice format `arrow`, as well as the newest format,
currently named `experimental`. This should clearly show how all of our
data types are expressed in an Arrow schema.

Currently, due to lack of support for `map` and `pattern` columns in
VAST, those two data types are not covered.