Fix record batch lifetime sharing
This change fixes an issue that was rather hard to run into, and far from simple
to debug once you did. Consider this example code:

```c++
table_slice A = get_table_slice();
auto B = as_record_batch(A);
auto C = B->Slice(...);
A = {};
B = {};
do_something_with(C); // crashes before this change
```

If C shares any memory with B, the above example crashed before this change.
This happens because because only A and B shared ownership of A's underlying
buffer, but the internal buffers of B did not. Any operation that created a new
record batch C that sliced referenced B's internal buffers did thus break the
structural lifetime-sharing with the table slice A.

This commit changes the lifetime-sharing setup and moves it from B to its
internal buffers. This works seemlessly with the existing stream decoder when
passing Arrow Buffers to it that structurally share the lifetime of the table
slice's underlying chunk. To avoid a cyclic reference in the table slice causing
a memory leak this additionally decouples the table slice implementation's
lifetime from the table slice's lifetime.