Add a coroutine-enabling wrapper for `caf::result<T>`
### `co_result`

This is best explained using a before-after comparison:

*Before*:

```cpp
[self](T&...) -> caf::result<R> {
  A();
  self->request(d1, t, xs...).then(
    [](response_type& value) {
      B(value);
    },
    [](caf::error& err) {
      C(err);
    });
  D();
  return r;
}
```

*After*:

```cpp
co_lift([self](T...) noexcept -> co_result<R> {
  A();
  auto task = co_request_then(*self, d1, t, xs...);
  D();
  auto tag = co_await co_deliver(r);
  if (auto value = co_await task)
    B(*value);
  else
    C(value.error());
  co_return tag;
})
```

The main benefit is immediately obvious: We've got rid of callback hell.

There exist a few pitfalls with this that might be easy to run into:
- Using `self->request`, `self->make_response_promise`, and `self->delegate`
  within a `co_result` coroutine leads to undefined behavior and is inherently
  unsafe. Use `co_request_then(self, ...)`, `co_request_await(self, ...)`,
  `co_make_response_promise<T>(self)`, and `co_delegate(self, ...)` instead.
- After awaiting a response handle, `self->current_sender()` might return an
  unexpected value because the resumed coroutine behaves as if it was called
  from the response handle continuation.
- To deliver values early, i.e., on first suspend, use `co_deliver(...)`, which
  returns a tag type that later must be returned from the coroutine.
- CAF supports catching exceptions via `set_exception_handler`. Throwing
  exceptions from a `co_result<T>` coroutine will always bypass CAF's exception
  handler and call `std::terminate()`.

### `co_request_then`

The analog to `self.request(dest, timeout, args...).then(...)` for
`co_result<T>` coroutines.

When awaited suspends the coroutine, returns an expected value similar to
`caf::function_view`, and continues the coroutine execution in the response
handlers CAF runtime context.

### `co_request_await`

The analog to `self.request(dest, timeout, args...).await(...)` for
`co_result<T>` coroutines.

When awaited suspends the coroutine, returns an expected value similar to
`caf::function_view`, and continues the coroutine execution in the response
handlers CAF runtime context.

### `co_request_receive`

The analog to `self.request(dest, timeout, args...).receive(...)` for
`co_result<T>` coroutines.

When awaited does *not* suspend the coroutine, returns an expected value similar
to `caf::function_view`, and continues the coroutine execution in the the
current CAF runtime context immediately.

### `co_deliver`

The analog to returning a value from an actor behavior for `co_result<T>`
coroutines. The value is returned to the CAF runtime on the next suspend.

When awaited does *not* suspend the coroutine, returns an empty tag value that
*must* later be returned from the coroutine, and continues the coroutine
execution in the the current CAF runtime context immediately.

This exists only to solve a corner case in which the return value of the
coroutine is known before suspending it via `co_await co_request_*`. Use with
caution.

### `co_delegate`

The analog to `self.delegate(dest, args...)` for `co_result<T>` coroutines.

When awaited does *not* suspend the coroutine, returns an empty tag value that
*must* later be returned from the coroutine, and continues the coroutine
execution in the the current CAF runtime context immediately.

### `co_make_response_promise`

The analog to `self.make_response_promise<T>()` for `co_result<T>` coroutines.

When awaited does *not* suspend the coroutine, returns the response promise and
an empty tag value that *must* later be returned from the coroutine, and
continues the coroutine execution in the the current CAF runtime context
immediately.

It is the caller's responsiblity to fulfill the returned response promise from
the same actor context.