Support wildcards in field suffix resolution
This adds support for replacing any part of a field extractor with '*' to jump
over a section. E.g., a query `suricata.*.flow.pkts_toserver > 0` will likely
hit events of the types `suricata.alert` and `suricata.flow` types. The wildcard
only ever jumps over a single section of a key, i.e., `suricata.*.pkts_toserver`
and `suricata.*.*.pkts_toserver` will match different fields.

This is implemented generically in the `resolve_key_suffix` function in VAST's
type system, which makes the same syntax available to all places that use suffix
matching for field names. E.g., the `project` transform can now be configured to
keep all fields in `suricata.flow.flow` by specifying `suricata.flow.flow.*`,
rather than having to list all fields in it individually.

Note that it does not make sense to implement this for the `resolve_key`
function as well, because that is required to return either zero or one results
by design, because it returns an exact match.