Fix bug in meta index roundtrip test
This test used to query with the field extractor `.x` for a field of the
fully-qualified name `y.x`. This only functioned for the meta index, which up
until recently simply ran `fqn.ends_with(".x")` to check for candidates. Now
it uses `layout.resolve_key_suffix(".x", "y")` instead, which fails because the
key suffix starts with a dot separator. This really isn't an issue for the user,
as this query never resulted in any results anyways—they were thrown out during
the evaluation of the candidates returned from the meta index.