Change the aggregate algorithm to make two passes
This is a rewrite of the aggregate transform step and its underlying algorithm
to have two passes, each of which can be disabled: An eager pass when adding
record batches, and a lazy pass when retrieving the results. Additionally, it
refactors the per-layout aggregation into a separate data structure that is
easier to maintain, and adds explanatory comments throughout the file.