Prefer CAF at `CAF_ROOT_DIR` over system CAF
This is a small bug fix to our build scaffolding, which preferred a CAF
installation in the default lookup paths over a manually specified installation
with `CAF_ROOT_DIR`. The solution is to not consider the specified location to
be a hint, but rather to be a path, and to hard-error if there is no CAF at the
specified location via `NO_DEFAULT_PATH`.