Fix race condition in CI for annotated tag pushes
If a dependent job runs git-describe to download an artifact uploaded from a
preceding job, we may run into race conditions because an annotated tag might've
been pushed in the meantime. This avoids that issue by only calling git-describe
once in the beginning, and then sharing that as a job output with all the other
jobs.