Improve aggregate performance massively
This change massively improves the performance of the aggregate transform by
getting rid of calls into Arrow Compute: While nice in theory, the overhead of
calling into the kernels was >80% of the time spent aggregating in benchmarks
with a median slice size of 3.

The new implementation does a lot of template trickery to allow for flexibly
working with scalars directly, which allows for keeping an accumulator per
columns and bucket that can simply be updated for every slice in the record
batch that fits into the same bucket.