Return kind of concrete types from `type::name()`
As a VAST developer using the type system, I would like to have a consistent and
meaningful interpretation of the type "name". There is currently a mismatch on
the query side and the implementation internally. For example, :timestamp has
the name "timestamp" whereas :time has no name. Given that there is one global
type namespace, this is an inconsistency in the model.

This change fixes that inconsistency throughout the type system.