Address review feedback
This removes the `type::has_name` function that caused confusion in review
comments, and instead introduces an intentionally private function that checks
whether a type has a custom name. This function is named `is_alias` and comes
with an exclaimer that it is intended to stay private because code using it
likely uses the wrong approach when interacting with the type system.

Additionally, this names some repetitive string literal constants as `T::kind`
for all `concrete_type T`.