Fix undefined behavior in field extractor
Using our current generator implementation with a non-const lvalue-reference to
the yielded valeus only works by chance. There is no guarantee that the
reference points to valid memory.

Sadly there's no way to guard against this behavior with the current
implementation; we may want to look at P2529 for inspritation on how to improve
it.