Correctly terminate the explore command
This commit fixes 2 issues that prevented the exporter for the initial
query from exiting.
When the explorer received the requested number of results for the
initial query it did not communicate this to the exporter, causing the
latter to wait for a follow-up request indefinitely.
When the `max-events-query * max-events-context` was lower than
`max-events` the sink would wait for additional results but the `explorer`
would only wait to get terminated.