Remove the concrete `none_type`
The `none_type` is a blast from the past: It originated in the old type system
that required an explicit variant member for signaling the empty type, which is
not a requirement for the new type system. This will allow for removing
`arrow::NullType` in VAST, which was—until now—required for symmetry in generic
code like the custom sum type access for `arrow::Array` and `arrow::DataType`.