Rename `meta-index` to `catalog`
We've long been searching for a new name for the meta index, as there exist
already multiple things named index inside VAST. An internal vote resulted in
the name `catalog` winning.

This change also affects a keys in the output of `vast status` and in certain
metrics, and is thus a breaking change.