Expose nested type aliases
The new function `type::aliases` returns all aliases of a type, from outermost
to innermost. Essentially, this iterates over the internal metadata structure of
a type and yields a slice of FlatBuffers table for every named metadata layer.

This is useful when wanting to write code that accepts both a type and any of
its aliases, e.g., as a key of a map for a custom index configuration, or when
tracking occurences of types in the catalog.