Give erase requests a high priority
In large-scale deployments we observed the disk monitor erase request to
return a timeout error, cancelling the ongoing request. This is a
two-fold bug fix: First, we must not use a timeout for process-internal
actor communication, especially not for such complicated nested loops
that when partially executed and then never resumed leave the actor in
an undefined state. Second, we must treat erase requests with a high
priority because they should never be upheld by queued requests on the
read or write path.