Fix race condition on timeout of query actor
A query gets removed from the list of 'pending' queries
in the index when the corresponding exporter actor goes
DOWN.

However, when the meta index lookup takes long enough the
exporter can time out before the query is added to 'pending'.
We did not check for this condition and thus would add
an already-dead query to 'pending', which of course would
never be able to progress afterwards.