Fix heap-use-after-free in aggregate group-by keys
It's not safe to store a `data_view` pointing into record batches whose
underlying data may no longer be around after calls to `aggregation::add` when
we want to access them in `aggregation::finish` later on. So now we store `data`
instead, and to avoid having to materialize as often we instead use a robin map
with support for heterogenous lookup from views.