Treat list options in env variables consistently
Before this change, there were four different ways to handle environment
variables for list options: Either through a `caf::config_option_set`, manually
split at a colon or a comma, or through the `data` parser.

Now, all list options are handled consistently, and can be specified either as
single values or as comma-separated lists that support proper escaping.

For the `VAST_SCHEMA_DIRS` and `VAST_PLUGIN_DIRS` option this is a subtle
breaking change, as they no longer can be specified with a colon-seperator, and
commas in the specified directories must now be escaped.