Fix direct value return after response promise has been created
When evaluating a query in the passive partition some code
paths would return a value after a response promise had been
generated, leading to a deadlock with the new scheduler
algorithm.

This case has become more common with the change to multiple
active partitions at once, because a query always goes to all
active partitions which are less likely to have a field
relevant to the query if they only contain events of a single type.