Don't prune string fields in the catalog
This removes an optimization from the catalog which would fold
all accesses to string fields into a single access to the
type-level string synopsis.

While efficient, the optimization was not correct in all cases.
For example, a query like

    http.hostname == "foo" || dns.rrname == "foo"

would be reduced to just `http.hostname == "foo"`, which would
return false for a partition that contains only `suricata.dns`
events regardless whether or not the second half of the disjunction
would have been true.