Fix query pruning in catalog
This several issues with the query pruner which was aggressively
reducing string lookups in compound statements using the fact that
all strings in a partition share the same bloom filter.

First, the relational operator was ignored leading to invalid
transformations like

    http.hostname == "foo" || http.hostname != "foo"

being "optimized" to just `http.hostname == "foo"` which is clearly
not the same.

Second, a query like

    http.hostname == "foo" || dns.rrname == "foo"

would also be reduced to just `http.hostname == "foo"`, which
returns false for a partition that contains only suricata.dns
events regardless whether or not the second half of the disjunction
would produced a candidate match.

Finally, with the introduction of configurable false-positive rates
for individual fields in VAST 2.0, even when applying this optimization
correctly it may lead to a lower-precision type level synopsis being
used for the candidate check.