Compress in-memory slices with Zstd
In a repeated test with 8'557'668 ingested Suricata events enabling the
compression had a negligible effect on import speed (under 0.5%), but reduced
the size of the archive and also the in-memory size of the loaded slices by
roughly the same amount.

Here are the sizes of the database for three scenarios: Uncompressed, Zstd with
the default compression level, and Zstd with the max compression level (50%
slower on important, so irrelevant in the grand scheme of things).

```
❯ du -csh vast.db.uncompressed/{index,archive}
667M	vast.db.uncompressed/index
1.8G	vast.db.uncompressed/archive
2.5G	total

❯ du -csh vast.db.zstd.default/{index,archive}
668M	vast.db.zstd.default/index
426M	vast.db.zstd.default/archive
1.1G	total

❯ du -csh vast.db.zstd.max/{index,archive}
675M	vast.db.zstd.max/index
361M	vast.db.zstd.max/archive
1.0G	total
```