Write homogenous partitions from the partition transformer
This changes the partition transformer to write not just
one output partition but rather a separate partition for
every output type.

In general the logic is "all-or-nothing", ie. the partition
transformer returns an error if a single output partition
fails to be persisted to disk.

This has a number of downstream implications, in particular
the index now also returns a vector of new partition infos
and the catalog has been expanded with a few bulk handlers.