Deduplicate explore hits by hash instead of ID
The explore command can fire of multiple sub-queries after initial
hits were produced for the given expression. When these sub-queries
produce overlapping result sets we need to make sure that we don't
present any events multiple times to the user. We achieve this by
maintainng a set of already forwarded events in the explorer actor.

Before this change that set consisted of event ids, relying on the
fact that those are globally unique. With this change the set to
contain event hashes instead. That means we have to produce a hash
of every result candidate, which is less efficient comapred to the
ids approach, but it shouldn't increase the delay noticeably for
the case where we only produce a small number of intermediate
results.

Changing the overall approach to wait for the initial query to
complete and generating a single big expression from the intermediate
results should be more efficient in general, but the change at hand
is less invasive.