Remove workaround for old combined layouts
This is a nuisance, and should much rather be hardcoded in old combined layouts
rather than being a part of the extractor resolution logic. It's a hard call to
make: On the one hand, implementing it as part of the extractor resolution is
easy, but on the other hand it also makes it impossible to correctly handle
wildcards. The proposed workaround is to instead use some heuristic to unflatten
the combined layout instead.