Mark erased partitions as such in the query queue
We don't remove the partition from the queue directly because the
query API requires clients to keep track of the number of candidate
partitions. Removing the partition from the queue would require us
to update the partition counters in the query states but the client
would go out of sync. That would require the index to deal with a
few complicated corner cases.
A partition that failed to load but was not erased can now be logged
with a proper warning thanks to this addition.