Respect the default fp-rate setting
VAST v2.0 introduced a new `vast.index` configuration subsection, and with that
came a bug we did not catch during testing: The defaults for the false positive
rate set under `vast.catalog-fp-rate` were not picked up at all. This fixes that
oversight, and moves the option to `vast.index.default-fp-rate` so it's right
where the other options are.