Relax check for aggregate transform steps
While technically the assertions this removes was correct, we have a fair amount
of transform steps that do not correctly label themselves an aggregate step.
Adding this assertion broke things that worked before, so we just leave it be
for now and are kind to the user. Since we plan to rework the transformation
engine in the medium term this is all rather irrelevant.