Support environment variables for plugin options
This is best explained on an example:

```
VAST_PLUGINS__EXAMPLE_ANALYZER__MAX_EVENTS=10 vast start
```

This will now correctly set the `max-events` option of the `example-analyzer`
plugin to 10. This did not work correctly before, because it referred to the
invalid option `vast.plugins.example-analyzer.max-events` rather than the
valid option `plugins.example-analyzer.max-events`.