Flush index state after partition transforms
In cases where the database only gets updated from partition transforms the
index state file can go heavily out of sync. This becomes a problem in case of
unexpected termination when the state can't get persisted, effectively resulting
in data loss.

This change adds calls to the persist function for that state in appropriate
places so that the problematic situation is prevented moving forward.