Defer slice serialization for construction from Arrow
This is an important optimization that causes us to persist slices to their IPC
backing at a later point in time only, i.e., when actually sending them across
the wire. This massively speeds up the performance of the Feather and Parquet
stores in most cases, although they're not quite comparable to the Segment store
yet because of some further unnecessary materializations.