Use a partition-local ID space for new partitions
This brings the query performance for Parquet and Feather stores up to the same
level as the as the segment store, because it allows them to use the selection
bitmap returned from the dense index lookup to reduce the number of rows
required to evaluate.

This is a bit of a prototype: All integration tests pass with all store
backends, which is what I've used for testing locally, but I did not adapt the
unit tests yet to the changes and they yet fail to build.

In addition, this also unveiled a shutdown issue with the new passive stores
that I do not understand yet, but that we need to look into: Shutdown takes a
long while even after queries completed. This is most noticeable with `vast
--node export --max-events=1 json`, which yields its one result almost instantly
but then takes seconds to shut down entirely. A similar behavior exists for
`vast rebuild`, which is terribly slow with the new stores.