Change expression evaluation to be column-oriented
This changes expression evaluation to primarily iterate over columns rather than
rows, which should in theory increase the number of cache hits and make it
easier to operate on record batches directly in the future.

The next step for optimizing this would be to extract the functionality of
`evaluate` and making it operate on a whole column at once rather than passing
in the individual fields.