Implement a flatbuffer container class to hold excess segments
This implements a new `SegmentedFile` flatbuffer as well as
a new `flatbuffer_container` utility class that uses this
flatbuffer to provide access to multiple flatbuffers stored
in the same file.

This is similar in spirit to the size-prefixed flatbuffers
provided by upstream, but allows random access into the
contained flatbuffers which will become more relevant when
we use the same approach for storing dense indices in partitions.

The segment is updated to optionally allow an implementation based
on flatbuffer containers, which should allow us to store segments
that hold more than 2GiB worth of table slice data.