Reset entire active partition state on decommission
This change fixes a bug for repeated calls to `flush` or calls to flush with a
very low `vast.active-partition-timeout` that was caused by us not resetting
the entire state the index held for the active partition. Flushing twice with no
additional events added in between should no longer cause errors, but rather
return (almost) immediately and do nothing.