Detach the catalog actor from CAF's thread pool
Running the catalog as a detached actor improved export performance under high
load for me by ~5% on my local machine when spawning 500 queries in parallel
with a ~100 GB database. The exact improvement depends on the workload, so YMMV.

The catalog is usually our busiest component, so giving it its own thread seemed
like a good fit. We should measure the effects of this change in a large-scale
deployment to find out whether we want to make it the default.