(WIP) Make sure terminated queries are removed from the backlog
When the index receives a new query it monitors the client so it
can clean up related state in case said client quits early.
Unfortunately this mechanism doesn't work in case the client
terminates before the index starts monitoring. In that case the
index-side state never gets cleaned and the queries remain in a
pending state indefinitely.

This commit introduces an extra roundtrip from the index to the
client. In case that new request times out we clean out the
internal state, and effectively cover the blind window.

ATTENTION:
It seems that request responses from scoped/blocking actors don't
get routed back to the requester correctly, which is a problem for
the index unit test.

TODO:
The query-processor / counter don't work yet.