Fix crash for aggregation functions returning lists
This fixes a crash for aggregation functions with different input and output
types, which just happened to work most of the time because the memory access
into the storage of an Arrow ListArray works transparently, so we passed the
unit tests. The additional assertions ensure that this works as expected.