Ignore native plugins when loaded explicitly
Native plugins are built-ins: They are always loaded by design. This change
makes it so we don't error if users specify them explicitly in their
configuration, but rather just silently ignore them. This makes the change of
the summarize plugin to be a native plugin no longer a breaking configuration
change.