Properly indicate failure in the rebuild command
The rebuild command triggered an assertion on failure in debug modes when
splitting old heterogeneous partitions, and it turned out that we did not
properly handle the failure case. This changes the rebuild command to show an
error if it did not succeed, detailing to the user what went wrong.