Make CI pass the correct version to Nix builds
This avoids a race condition: GitHub's steps run asynchronously, and there's no
guarantee that the output of `git-describe` is consistent between steps, because
an annotated tag may have been pushed in the mean time. For the v2.1 release,
this caused the version number of the binary attached to the release to be off
by one.