Remove the top-level systemd directory
Alongside the new option to create Debian packages from our build scaffolding we
mvoed the systemd unit file templates to `vast/services/systemd`. The old
top-level `systemd` directory remained as the contained configuration file was
used by the top-level Dockerfile. This commit cleans that up by removing the
directory and switching the Dockerfile to use environment variables rather than
a copied configuration file, which was not an option at the time the Dockerfile
was created.