Split install into Runtime and Development components
This is a fix for a regression we purposefully introduced when switching to
building Debian packages in CI which caused created packages to unnecessarily
include development dependencies for the static binary. Essentially, calls to
`install(...)` in CMake now require a `COMPONENT Runtime` or `COMPONENT
Development` categorization, and we make use of then when creating packages, or
users can choose to install just parts the runtime-relevant parts of VAST when
building from source by running `cmake --install path/to/build --component
Runtime`.