Add a healthcheck to the tenzir/vast Docker image
The health check is primarily useful container orchestration, where it allows
for detecting (1) when a VAST server can first be connected to and (2) when a
VAST server is unresponsive.

For example, in a Docker Compose file services running `vast import` can depend
on a `vast start` service being healthy instead of just running to avoid failing
to connect because the service started, but did not listen for incoming
connections yet.