Avoid code coverage blocking CI runs on master
For the VAST workflow we defined a concurrency section that ensures runs on the
master branch happen sequentially, avoiding all possible races in artifact
uploads or unnecessary cache invalidations. The code coverage run takes up to
two hours regularly, so this change splits it out into another workflow with a
different concurrency section.